Externalize event builder class. Pass mapping as argument of event builder and Commissioning event. Fix calo and tracker mapping for programs / tests. Mock calibration works only for calorimeter hits and can been with flvisualize.
