+++ companions/fecom/resources/config/default_mapping_tracker.csv
@@ -1,0 +1,37 @@
+#SlotID CellLayer CellRow AnodicChannel BotCathodeChannel TopCathodeChannel
+11 0 0 0 1 2
+11 1 0 4 3 8
+11 2 0 6 5 14
+11 3 0 10 7 20
+11 4 0 12 9 26
+11 5 0 16 11 32
+11 6 0 18 13 38
+11 7 0 22 15 44
+11 8 0 24 17 50
+11 0 1 28 19 56
+11 1 1 30 21 62
+11 2 1 34 23 68
+11 3 1 36 25 74
+11 4 1 40 27 80
+11 5 1 42 29 86
+11 6 1 46 31 92
+11 7 1 48 33 98
+11 8 1 52 35 104
+11 0 2 54 55 37
+11 1 2 58 57 39
+11 2 2 60 59 41
+11 3 2 64 61 43
+11 4 2 66 63 45
+11 5 2 70 65 47
+11 6 2 72 67 49
+11 7 2 76 69 51
+11 8 2 78 71 53
+11 0 3 82 73 91
+11 1 3 84 75 93
+11 2 3 88 77 95
+11 3 3 90 79 97
+11 4 3 94 81 99
+11 5 3 96 83 101
+11 6 3 100 85 103
+11 7 3 102 87 105
+11 8 3 106 89 107
