+++ companions/fecom/resources/config/mapping_calo.csv
@@ -1,27 +1,27 @@
 #SlotID Channel Column Row
-15 0 0 0
-15 1 0 1
-15 2 0 2
-15 3 0 3
-15 4 0 4
-15 5 0 5
-15 6 0 6
-15 7 0 7
-15 8 0 8
-15 9 0 9
-15 10 0 10
-15 11 0 11
-15 12 0 12
-16 0 1 0
-16 1 1 1
-16 2 1 2
-16 3 1 3
-16 4 1 4
-16 5 1 5
-16 6 1 6
-16 7 1 7
-16 8 1 8
-16 9 1 9
-16 10 1 10
-16 11 1 11
-16 12 1 12
+2 0 0 0
+2 1 0 1
+2 2 0 2
+2 3 0 3
+2 4 0 4
+2 5 0 5
+2 6 0 6
+2 7 0 7
+2 8 0 8
+2 9 0 9
+2 10 0 10
+2 11 0 11
+2 12 0 12
+3 0 1 0
+3 1 1 1
+3 2 1 2
+3 3 1 3
+3 4 1 4
+3 5 1 5
+3 6 1 6
+3 7 1 7
+3 8 1 8
+3 9 1 9
+3 10 1 10
+3 11 1 11
+3 12 1 12
