+++ companions/fecom/resources/config/mapping_tracker.csv
@@ -1,37 +1,10 @@
 #SlotID CellLayer CellRow AnodicChannel BotCathodeChannel TopCathodeChannel
-0 0 0 0 1 2
-0 1 0 4 3 8
-0 2 0 6 5 14
-0 3 0 10 7 20
-0 4 0 12 9 26
-0 5 0 16 11 32
-0 6 0 18 13 38
-0 7 0 22 15 44
-0 8 0 24 17 50
-0 0 1 28 19 56
-0 1 1 30 21 62
-0 2 1 34 23 68
-0 3 1 36 25 74
-0 4 1 40 27 80
-0 5 1 42 29 86
-0 6 1 46 31 92
-0 7 1 48 33 98
-0 8 1 52 35 104
-0 0 2 54 55 37
-0 1 2 58 57 39
-0 2 2 60 59 41
-0 3 2 64 61 43
-0 4 2 66 63 45
-0 5 2 70 65 47
-0 6 2 72 67 49
-0 7 2 76 69 51
-0 8 2 78 71 53
-0 0 3 82 73 91
-0 1 3 84 75 93
-0 2 3 88 77 95
-0 3 3 90 79 97
-0 4 3 94 81 99
-0 5 3 96 83 101
-0 6 3 100 85 103
-0 7 3 102 87 105
-0 8 3 106 89 107
+11 8 4 54 89 37
+11 7 4 58 87 39
+11 6 4 60 85 41
+11 5 4 64 83 43
+11 4 4 66 81 45
+11 3 4 70 79 47
+11 2 4 72 77 49
+11 1 4 76 75 51
+11 0 4 78 73 53
