+++ companions/fecom/src/libs/libfecom/fecom/tracker_channel_hit.hpp
@@ -27,16 +27,16 @@ namespace fecom {
     /// Predicate to find a tracker channel hit by channel
     struct find_by_channel {
       find_by_channel(const uint16_t & feast_id_,
-		      const uint16_t & channel_id_) : feast_id(feast_id_),
-						      channel_id(channel_id_) {}
+                      const uint16_t & channel_id_) : feast_id(feast_id_),
+                                                      channel_id(channel_id_) {}
 
       bool operator()(const tracker_channel_hit & tracker_channel_hit_) {
-	uint16_t input_feast_id = tracker_channel_hit_.electronic_id.get(tracker_constants::FEAST_INDEX);
-	uint16_t input_channel_id = tracker_channel_hit_.electronic_id.get(tracker_constants::CHANNEL_INDEX);
-	bool finded = false;
-	if (feast_id == input_feast_id && channel_id == input_channel_id) finded = true;
-	else finded = false;
-	return finded;
+        uint16_t input_feast_id = tracker_channel_hit_.electronic_id.get(tracker_constants::FEAST_INDEX);
+        uint16_t input_channel_id = tracker_channel_hit_.electronic_id.get(tracker_constants::CHANNEL_INDEX);
+        bool finded = false;
+        if (feast_id == input_feast_id && channel_id == input_channel_id) finded = true;
+        else finded = false;
+        return finded;
       }
 
       uint16_t feast_id;
@@ -46,21 +46,21 @@ namespace fecom {
     /// Predicate to find a tracker channel hit by timestamp
     struct find_by_timestamp {
       find_by_timestamp(const uint16_t & feast_id_,
-			const uint16_t & channel_id_,
-			const std::string & timestamp_type_) : feast_id(feast_id_),
-							       channel_id(channel_id_),
-							       timestamp_type(timestamp_type_) {}
+                        const uint16_t & channel_id_,
+                        const std::string & timestamp_type_) : feast_id(feast_id_),
+                                                               channel_id(channel_id_),
+                                                               timestamp_type(timestamp_type_) {}
 
       bool operator()(const tracker_channel_hit & tracker_channel_hit_) {
-	uint16_t input_feast_id = tracker_channel_hit_.electronic_id.get(tracker_constants::FEAST_INDEX);
-	uint16_t input_channel_id = tracker_channel_hit_.electronic_id.get(tracker_constants::CHANNEL_INDEX);
-	std::string input_timestamp_type = tracker_channel_hit_.timestamp_type;
-	bool finded = false;
-	if (feast_id == input_feast_id
-	    && channel_id == input_channel_id
-	    && timestamp_type == input_timestamp_type) finded = true;
-	else finded = false;
-	return finded;
+        uint16_t input_feast_id = tracker_channel_hit_.electronic_id.get(tracker_constants::FEAST_INDEX);
+        uint16_t input_channel_id = tracker_channel_hit_.electronic_id.get(tracker_constants::CHANNEL_INDEX);
+        std::string input_timestamp_type = tracker_channel_hit_.timestamp_type;
+        bool finded = false;
+        if (feast_id == input_feast_id
+            && channel_id == input_channel_id
+            && timestamp_type == input_timestamp_type) finded = true;
+        else finded = false;
+        return finded;
       }
 
       uint16_t feast_id;
@@ -77,8 +77,8 @@ namespace fecom {
     /// Check if the tracker
     virtual bool is_valid() const;
 
-    ///
-    virtual double get_timestamp() const;
+    /// Return the integer timestamp
+    virtual double get_timestamp() const override;
 
     /// Reset
     virtual void reset();
@@ -100,7 +100,7 @@ namespace fecom {
   public:
 
     // Management :
-    mutable bool associated = false; ///< Flag to know if the tracker channel is already asociate in a tracker_hit
+    mutable bool associated = false; ///< Flag to know if the tracker channel is already associated to a tracker_hit
 
     // Config :
 
@@ -108,7 +108,7 @@ namespace fecom {
 
     // Data :
     uint64_t timestamp_value;      ///< Value of the timestamp
-    double timestamp_time_ns;      ///< Value of the timestamp in ns
+    double timestamp_time_ns;      ///< Value of the timestamp in ns (redunduncy from DAQ)
 
     DATATOOLS_SERIALIZATION_DECLARATION()
